### 2020-05-27 #2264 @whedon
[PRE REVIEW]: Exploring meaningful visual effects and quantities of interest from dynamic models through dynamac
**Submitting author:** @andyphilips (<a href="http://orcid.org/0000-0003-4201-1085">Soren Jordan</a>)
**Repository:** <a href="https://github.com/andyphilips/dynamac" target ="_blank">https://github.com/andyphilips/dynamac</a>
**Version:** v0.1.10
