### 2020-05-27 #2264 @danielskatz
Can you also say a little in the paper about what types of research (in what fields, for example) would use this software, and for what purpose?
