### 2020-05-27 #2264 @danielskatz
some of this might be in the conclusions, but it should really be in the introduction
