### 2020-05-27 #2264 @danielskatz
You can then enter ```@whedon generate pdf``` as a new comment here to regenerate the PDF.
