### 2020-05-29 d4df5f5
added details to intro
