+++ JOSS/paper.md
@@ -23,7 +23,9 @@ bibliography: jordan-philips.bib
 
 # Summary 
 
-dynamac [-@dynamacCRAN] uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. Such dynamic models appear in a variety of fields such as environmental science [@sotte2019;@sotte2020;@jcp,@ijerph], economics [@espr;@te], and political science [@trumpy]. dynamac implements these tools in a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. The procedure is as follows. First, the parameters are estimated using an ARDL model. Next, the parameters used to draw $s$ simulations from a multivariate normal distribution, with mean $\hat{\beta}$ and variance obtained from the estimated variance-covariance matrix. Stable predicted values for $y$ using the $s$ simulations of $\hat{\beta}$ are then created, using starting values of the independent variables (usually means for continuous variables or modes for categorical ones). Users can choose from either expected values or---by incorporating fundamental model uncertainty---predicted values. Next, at a defined time period $t$, one of the independent variables is "shocked" by an amount determined by the researcher, and the effect on the dependent variable is observed across the resulting time periods. To summarize, models are estimated and counterfactual shocks are simulated through `dynardl()`. Then post-estimation dynamic inferences can be uncovered through the new command `dynardl.simulation.plot()`, which has six plotting responses mapped to the six visualizations that we outlined below. 
+dynamac [-@dynamacCRAN] implements a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. Further, dynamac uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. These simulation techniques bring a wider set of inferences from complex models to less methodologically sophisticated users in fields as diverse as environmental science [@sotte2019;@sotte2020;@jcp;@ijerph], epidemiology [@covid], economics [@espr;@te], political science [@trumpy], and beyond. Users can make dynamic inferences from their models, including measures of uncertainty, without the need for any formulae or algebraic solutions.
+
+The stochastic simulation procedure is as follows. First, the parameters are estimated using an ARDL model. Next, the parameters used to draw $s$ simulations from a multivariate normal distribution, with mean $\hat{\beta}$ and variance obtained from the estimated variance-covariance matrix. Stable predicted values for $y$ using the $s$ simulations of $\hat{\beta}$ are then created, using starting values of the independent variables (usually means for continuous variables or modes for categorical ones). Users can choose from either expected values or---by incorporating fundamental model uncertainty---predicted values. Next, at a defined time period $t$, one of the independent variables is "shocked" by an amount determined by the researcher, and the effect on the dependent variable is observed across the resulting time periods. To summarize, models are estimated and counterfactual shocks are simulated through `dynardl()`. Then post-estimation dynamic inferences can be uncovered through the new command `dynardl.simulation.plot()`, which has six plotting responses mapped to the six visualizations that we outline below. 
 
[…]
