+++ JOSS/jordan-philips.bib
@@ -125,6 +125,18 @@
 	doi = {10.1016/j.jclepro.2019.119533}
 }
 
+@article{covid,
+	Author = {Sarkodie, Samuel Asumau and Owusu, Phebe Asantewaa},
+	Journal = {Heliyon},
+	volume = {6},
+	number = {4},
+	Pages = {},
+	Title = {Investigating the cases of novel coronavirus disease ({COVID}-19) in China using dynamic statistical techniques},
+	Year = {2020},
+	doi = {10.1016/j.heliyon.2020.e03747}
+}
+
+
 @article{sotte2020,
 	Author = {Sarkodie, Samuel Asumau and Adams, Samuel and Owusu, Phebe Asantewaa and Leirvik, Thomas and Ozturk, Ilhan},
 	Journal = {Science of the Total Environment},
