### 2020-06-02 #2264 @andyphilips
Thanks @danielskatz! Repository now has an Apache 2.0 license, and we updated the intro, as suggested. I just regenerated the PDF, but please let me know if there's anything else I need to do.
