### 2020-06-12 #2264 @labarba
👋  @KristinaRiemer — Could you have a look at this submission and let us know if you can handle it as editor?
Title — Exploring meaningful visual effects and quantities of interest from dynamic models through dynamac
Language — R
[Article proof](https://github.com/openjournals/joss-papers/blob/joss.02264/joss.02264/10.21105.joss.02264.pdf)
