### 2020-06-16 #2264 @KristinaRiemer
I'm not sure that I know enough about this statistical approach to be a good editor? Also, I don't know if the Stata version is included with this submission, but I have no experience with Stata.
