### 2020-06-18 #2264 @usethedata
While I'd love to help, the stats domain for this package and the things they're doing in R are pretty far outside my areas of expertise.
