### 2020-06-18 #2264 @arfon
@Kevin-Mattheus-Moerman - this might be a good first submission for @sbenthall who is currently in the process of joining the JOSS editorial team.
