### 2020-06-19 #2264 @arfon
> @whedon invite @sbenthall as editor

I've not completed the onboarding of @sbenthall quite yet (so that invite didn't work) but he should be on the team within the next couple of days.
