### 2020-06-30 #2264 @RomainAzais
> @RomainAzais would you be wiling to review this paper?

Hello @sbenthall
I'm sorry, but, at the moment, I'm not available to review this submission, and I have no experience with Stata. I hope to be able to help next time.
