### 2020-07-04 #2264 @sbenthall
To clarify: the software under review in this issue is written in R.
