### 2020-07-07 #2264 @kauedesousa
Yes @sbenthall but I only provide my feedback in the first week of August. Is that OK?
