### 2020-07-13 #2264 @stulacy
I'll be able to review the R code but I don't have a Stata licence to review that software. Would that be an issue?
