### 2020-07-16 #2264 @sbenthall
@stulacy That's a good question. I'm new here. Let me ask ... @xuanxu what happens if a submission has a proprietary dependency?
