### 2020-07-17 #2264 @xuanxu
@sbenthall In this case I think all the review can be done using R, but to accept the claim that the software works in Stata a reviewer should have access to Stata at least to test the package installation guide from the README file. Is it possible for the authors to grant the reviewers temporary access to a Stata instance?

Maybe some @openjournals/joss-eics have any other idea.
