### 2020-07-17 #2264 @andyphilips
Hi @sbenthall ....I think there may be some confusion. Just to clarify, the JOSS review is for the updated R version of this program, not the Stata version. The Stata version has already been reviewed/published elsewhere. I know that both the R and Stata programs are contained in the same branch (we did this to try and keep everything in one location); do I need to move one of them?
