### 2020-07-22 #2264 @dormezil
Yes, I am willing to review Exploring meaningful visual effects and
quantities of interest from dynamic models through dynamac.

Best Regards,

Stevens Dormezil
Data Scientist
B. ‪(561) 425-9037‬

On Wed, Jul 22, 2020 at 3:00 PM Sebastian Benthall <notifications@github.com>
wrote:

> @dormezil <https://github.com/dormezil> Would you be willing to review
> this article? Please note the scope of the review is the R
> implementation only.
>
