### 2020-07-28 #2528 @whedon
[REVIEW]: Exploring meaningful visual effects and quantities of interest from dynamic models through dynamac
**Submitting author:** @andyphilips (<a href="http://orcid.org/0000-0003-4201-1085">Soren Jordan</a>)
**Repository:** <a href="https://github.com/andyphilips/dynamac" target ="_blank">https://github.com/andyphilips/dynamac</a>
**Version:** 0.1.11
