### 2020-08-05 #2528 @stulacy
This was an interesting read and the software looks to be useful in helping extract measures of interest from ARDL models using simulation. My comments are:

- Can you briefly summarise in the paper how the `dynamac` package compares to other available software, such as `ARDL` or `dlagM`, which are both also available on CRAN
- Figure 1 in the paper shows all the 6 available plots that the `dynardl.simulation.plot` function produces. However, 3 of these (top left, top middle, bottom middle) all look identical. Would it be possible to use an example model and scenario where these 3 plots differ to better identify the purpose of each?
- Minor typo in the second sentence of the second paragraph, "Next, the parameters [are] used to draw...."
