### 2020-09-09 #2528 @dormezil
The software works and is useful for working with ARDL models. There is a typo in the 2nd paragraph, "Next, the parameters [are] used ...."
