+++ joss-doc/jordan-philips.bib
@@ -35,6 +35,29 @@
 	Year = {2020}
 }
 
+@manual{dynlmCRAN,
+	Author = {Zeileis, Achim},
+	Note = {R package version 0.3-6},
+	Title = {dynlm: Dynamic Linear Models and Time Series Regression},
+	Url = {https://CRAN.R-project.org/package=dynlm},
+	Year = {2019}
+}
+
+@manual{ardlCRAN,
+	Author = {Natsiopoulos, Kleanthis and Tzeremes, Nickolaos},
+	Note = {R package version 0.1.0},
+	Title = {{ARDL}: {ARDL}, {ECM} and Bounds-Test for Cointegration},
+	Url = {https://CRAN.R-project.org/package=ARDL},
+	Year = {2020}
+}
+
+@manual{dlagMCRAN,
+	Author = {Demirhan, Haydar},
+	Note = {R package version 1.1.4},
+	Title = {{dLagM}: Time Series Regression Models with Distributed Lag Models},
+	Url = {https://CRAN.R-project.org/package=dLagM},
+	Year = {2020}
+}
 
 @article{jordan2018dynamic,
 	Author = {Jordan, Soren and Philips, Andrew Q},
@@ -124,6 +147,18 @@
 	Year = {2020},
 	doi = {10.1016/j.jclepro.2019.119533}
 }
+
+@article{covid,
+	Author = {Sarkodie, Samuel Asumau and Owusu, Phebe Asantewaa},
+	Journal = {Heliyon},
+	volume = {6},
+	number = {4},
+	Pages = {},
+	Title = {Investigating the cases of novel coronavirus disease ({COVID}-19) in China using dynamic statistical techniques},
+	Year = {2020},
+	doi = {10.1016/j.heliyon.2020.e03747}
+}
+
 
 @article{sotte2020,
 	Author = {Sarkodie, Samuel Asumau and Adams, Samuel and Owusu, Phebe Asantewaa and Leirvik, Thomas and Ozturk, Ilhan},
