+++ JOSS/paper.md
@@ -11,7 +11,7 @@ affiliations:
     name: Assistant Professor, Department of Political Science, Auburn University 
   - index: 2 
     name: Assistant Professor, Department of Political Science, University of Colorado Boulder 
-date: 9 April 2020 
+date: 17 September 2020 
 tags: 
   - time series
   - methodology
@@ -23,9 +23,9 @@ bibliography: jordan-philips.bib
 
 # Summary 
 
-dynamac [-@dynamacCRAN] implements a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. Further, dynamac uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. These simulation techniques bring a wider set of inferences from complex models to less methodologically sophisticated users in fields as diverse as environmental science [@sotte2019;@sotte2020;@jcp;@ijerph], epidemiology [@covid], economics [@espr;@te], political science [@trumpy], and beyond. Users can make dynamic inferences from their models, including measures of uncertainty, without the need for any formulae or algebraic solutions.
+dynamac [-@dynamacCRAN] implements a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. Further, dynamac uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. These simulation techniques bring a wider set of inferences from complex models to less methodologically sophisticated users in fields as diverse as environmental science [@sotte2019;@sotte2020;@jcp;@ijerph], epidemiology [@covid], economics [@espr;@te], political science [@trumpy], and beyond. Users can make dynamic inferences from their models, including measures of uncertainty, without the need for any formulae or algebraic solutions. While other R packages may estimate ARDL models (for instance dynlm [-@dynlmCRAN], which only provides model estimates; ardl [-@ardlCRAN], which uses dynlm for estimation but provides support for the bounds test for cointegration as well as automated lag selection; dlagM [-@dlagMCRAN], which allows n-step model forecasting), dynamac is unique in that it is the only R package that provides post-estimation diagnostics for autocorrelation in the residuals and is designed specifically for providing inferences through counterfactual simulation.
 
-The stochastic simulation procedure is as follows. First, the parameters are estimated using an ARDL model. Next, the parameters used to draw $s$ simulations from a multivariate normal distribution, with mean $\hat{\beta}$ and variance obtained from the estimated variance-covariance matrix. Stable predicted values for $y$ using the $s$ simulations of $\hat{\beta}$ are then created, using starting values of the independent variables (usually means for continuous variables or modes for categorical ones). Users can choose from either expected values or---by incorporating fundamental model uncertainty---predicted values. Next, at a defined time period $t$, one of the independent variables is "shocked" by an amount determined by the researcher, and the effect on the dependent variable is observed across the resulting time periods. To summarize, models are estimated and counterfactual shocks are simulated through `dynardl()`. Then post-estimation dynamic inferences can be uncovered through the new command `dynardl.simulation.plot()`, which has six plotting responses mapped to the six visualizations that we outline below. 
[…]
