+++ JOSS/paper.md
@@ -23,7 +23,7 @@ bibliography: jordan-philips.bib
 
 # Summary 
 
-dynamac [-@dynamacCRAN] implements a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. Further, dynamac uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. These simulation techniques bring a wider set of inferences from complex models to less methodologically sophisticated users in fields as diverse as environmental science [@sotte2019;@sotte2020;@jcp;@ijerph], epidemiology [@covid], economics [@espr;@te], political science [@trumpy], and beyond. Users can make dynamic inferences from their models, including measures of uncertainty, without the need for any formulae or algebraic solutions. While other R packages may estimate ARDL models (for instance dynlm [-@dynlmCRAN], which only provides model estimates; ardl [-@ardlCRAN], which uses dynlm for estimation but provides support for the bounds test for cointegration as well as automated lag selection; dlagM [-@dlagMCRAN], which allows n-step model forecasting), dynamac is unique in that it is the only R package that provides post-estimation diagnostics for autocorrelation in the residuals and is designed specifically for providing inferences through counterfactual simulation.
+dynamac [-@dynamacCRAN] implements a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. Further, dynamac uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. These simulation techniques bring a wider set of inferences from complex models to less methodologically sophisticated users in fields as diverse as environmental science [@sotte2019;@sotte2020;@jcp;@ijerph], epidemiology [@covid], economics [@espr;@te], political science [@trumpy], and beyond. Users can make dynamic inferences from their models, including measures of uncertainty, without the need for any formulae or algebraic solutions. While other R packages may estimate ARDL models (for instance dynlm [@dynlmCRAN], which only provides model estimates; ardl [@ardlCRAN], which uses dynlm for estimation but provides support for the bounds test for cointegration as well as automated lag selection; dlagM [@dlagMCRAN], which allows n-step model forecasting), dynamac is unique in that it is the only R package that provides post-estimation diagnostics for autocorrelation in the residuals and is designed specifically for providing inferences through counterfactual simulation.
 
[…]
