### 2020-09-18 #2528 @andyphilips
@sbenthall @dormezil @stulacy Thank you all for the thoughtful reviews. We outline our particular response and revisions below. We have also uploaded all relevant files to the repository.

@stulacy:

**Point 1.**
We appreciate the opportunity to differentiate dynamac from other R packages. @stulacy  is correct in that multiple packages estimate ARDL models: dynlm [-@dynlmCRAN] (which is just model estimation), ardl [-@ardlCRAN] (a front-end for dynlm that extends it to the bounds test for cointegration [which dynamac also provides]), and dlagM [-@dlagMCRAN] (which offers basic model forecasting but in a clunky ARDL(p, q) estimation framework). We credit these packages as alternatives for users while illuminating the main difference for dynamac, which really is the thrust of the paper: compelling visualizations produced by counterfactual simulation. To illustrate this difference, we've added an additional sentence to the first paragraph.

**Point 2**.
It's correct that three of the plots look very similar (and would look very similar in any example). However, they are more finely differentiated by serving different purposes. We try to clarify these separate purposes through the text. The first panel presents Y in levels. This is interesting for an obvious reason: if the researcher wanted to predict where Y would end as a result to a shock in X propagating through the system, s/he would just want the fitted values of Y.

But suppose the researcher instead wanted to know how much Y changed from its starting value, rather than its new level per se. This would require some calculation on the researcher's end. We aim to minimize this burden in the second panel, ``Changes from Y Mean Value,'' is identical to the first but subtracts out the mean value of Y. This would be especially helpful if the response series re-equilibrated to a new value as a result of the shock to the X series. Instead of having to calculate whether this re-equilibration occurred, subtracting out the mean of Y gives a measure of how much Y changed from its pre-shock equilibrium mean.

The fifth plot, ``Cumulative Change in Y Value,'' is different in its calculation. For the second panel, the fitted values are calculated within each period, with the mean of Y subtracted out. This leads to an estimate of the within-period change in Y from its mean value, with confidence intervals taken from the percentiles of the fitted values within each period of the simulation. The fifth plot tracks the simulations across all time periods; the plotted estimates are the cumulative changes in Y across all of the simulations up to the corresponding period. This will almost always be more conservative in the estimation of the confidence intervals than tracking the within-period changes (without considering the individual histories of the simulations to that time period), as in the second panel. It also more closely reflects the logic of the long-run multiplier it is meant to emulate: the cumulative change in Y that results from a shock to X (rather than just the within-period position of Y relative to its mean).

To clarify, we have now constructed a new example that illustrates the more conservative approach to confidence intervals between the second and fifth panels. In the text, we clarify this under the second and fifth bullet points:

``Deviations between the predicted and average values of $y_t$ (i.e., $\hat{y}_t - \bar{y}$). While this plot follows a response path identical to the plot above, if a shock dissipates and the series reverts to its mean, this is easier to observe if we do not have to subtract the stable starting value from the plot. If the series reverts to a value other than its mean, we would be especially interested in this new long-run mean in response to the shock. This is analogous to an impulse response function.''

``The cumulative response in $y_t$ to a shock in $x_t$. This helps us to understand what the "final" effect of a shock to an independent variable does to a dependent variable. An independent variable might first cause a positive response, but ultimately may be overwhelmed by a negative movement as time goes on. Unlike the other plots, this considers the cumulative histories in each individual simulation when plotting the response in $y$ over time, making the estimation of the confidence intervals more conservative. This is analogous to the traditional long-run effect of an ARDL model; however, it requires no analytical calculation.''

Finally, we'd like to note that even though the plots look similar, they are built to serve different logics and different approaches to calculating quantities of interest. They are also completely new and not found in any other package, so we wanted to provide a diverse set of approaches to practitioners without adjudicating for them which plots they should use. And we suspect that most of the time, users will choose one of the plot types to show (through dynardl.simulation.plot) rather than showing all of them, minimizing the redundancy.

**Point 3** (which @dormezil also carefully noted).

Thank you for your careful reading: we amended the typo!
