### 2020-09-22 #2528 @stulacy
Thank you for your detailed response. I am satisfied my points have been addressed.
