### 2020-09-24 #2528 @sbenthall
I have one comment on the paper: it might help the reader if you provide the ARDL equation early on.

To proceed with the editorial process, @andyphilips would you please:

- Do a final proofread of your paper (it will be much more difficult to change the paper after it is accepted.
- issue a new tagged release of the software (if changed)
- archive it (on Zenodo, figshare, or other)
- post the version number and archive DOI here
