### 2020-09-28 3de21bf
final JOSS paper submission; fixed command line error
