+++ JOSS/paper.md
@@ -40,7 +40,7 @@ This functionality vastly simplifies the production and interpretation of dynami
 In addition to the difficulties in calculating short- and long-run effects, inferences can also be obscured due to the complexity of the model. This is especially true as the models become more sophisticated (like including multiple lags and/or first-differences of dependent and independent variables). For short-run effects, increased model complexity encourages a focus on "boring" effects (the one-unit effect reported by the $\hat\beta$ in the regression output); for long-run effects, greater model complexity results in increasingly intractable closed-form estimates of the effects. Stochastic simulations can solve both of these problems by allowing for short-run dynamics beyond the traditional one-unit effect, and also by removing the need for analytic calculations when discussing long-run effects. 
 
 # Plotting functionality
-Specifically, dynamac implements six new visualizations. These can be presented individually using `dynardl.simulation.plot()`, or all together using the command `dynardl.all.plots()`. The resulting plots are:
+Specifically, dynamac implements six new visualizations. These can be presented individually using the command `dynardl.simulation.plot()`, or all together with `dynardl.all.plots()`. The resulting plots are:
 
 * The response path of the level of the dependent variable, given a change (a "shock") in a single independent variable at a single point in time [@williams2012but]. Note the value of this "shock" is user-defined, allowing for meaningful short-run inferences beyond the typical one-unit effect represented by a $\hat\beta$; for instance, a one standard deviation increase or decrease.
 * Deviations between the predicted and average values of $y_t$ (i.e., $\hat{y}_t - \bar{y}$). While this plot follows a response path identical to the plot above, if a shock dissipates and the series reverts to its mean, this is easier to observe if we do not have to subtract the stable starting value from the plot. If the series reverts to a value other than its mean, we would be especially interested in this new long-run mean in response to the shock. This is analogous to an impulse response function.
