+++ joss-doc/paper.md
@@ -1,63 +1,0 @@
---- 
-title: 'Exploring meaningful visual effects and quantities of interest from dynamic models through dynamac'  
-authors:
-  - name: Soren Jordan 
-    affiliation: 1 
-    orcid: 0000-0003-4201-1085 
-  - name: Andrew Q. Philips 
-    affiliation: 2 
-affiliations: 
-  - index: 1 
-    name: Assistant Professor, Department of Political Science, Auburn University 
-  - index: 2 
-    name: Assistant Professor, Department of Political Science, University of Colorado Boulder 
-date: 28 September 2020 
-tags: 
-  - time series
-  - methodology
-  - long-run effects
-  - simulations
-  - quantities of interest
-bibliography: jordan-philips.bib 
----
-
-# Summary 
-
-dynamac [-@dynamacCRAN] implements a cohesive ecosystem of estimating and interpreting autoregressive distributed lag (ARDL) models. Further, dynamac uses stochastic simulation techniques [@jordan2018cointegration;@jordan2018dynamic] to easily recover traditional quantities of interest, such as short- and long-run effects, even from complex dynamic specifications (such as the addition of multiple lags of dependent variables, or independent variables appearing in multiple lags or multiple forms like contemporaneous and lagged differences, or contemporaneous and lagged levels), including models with cointegration [@philips2016have]. These simulation techniques bring a wider set of inferences from complex models to users in fields as diverse as environmental science [@sotte2019;@sotte2020;@jcp;@ijerph], epidemiology [@covid], economics [@espr;@te], political science [@trumpy], and beyond. Users can make dynamic inferences from their models, including measures of uncertainty, without the need for any formulae or algebraic solutions. Although other R packages may estimate ARDL models (for instance dynlm [@dynlmCRAN], which only provides model estimates; ardl [@ardlCRAN], which uses dynlm for estimation but provides support for the bounds test for cointegration as well as automated lag selection; dlagM [@dlagMCRAN], which allows n-step model forecasting), dynamac is unique in that it is the only R package that provides post-estimation diagnostics for autocorrelation in the residuals and is designed specifically for providing inferences through counterfactual simulation.
-
-The stochastic simulation procedure is as follows. First, the parameters are estimated using an ARDL model, a very general form of which can be given as:
-\begin{align}\label{eq:ardlgeneral}
-\begin{split}
-	y_t,\Delta y_t =  &\alpha_0 + \delta T + \sum_{p=1}^P \phi_p y_{t-p} + \sum_{l_1 = 0}^{L_1} \theta_{1l}x_{1,t-l_1} + \cdots + \sum_{l_k = 0}^{L_k} \theta_{kl}x_{t-l_k} + \\
-	& \sum_{m=1}^M \alpha_m \Delta y_{t-m} + \sum_{q_1=0}^{Q_1} \beta_{1q_{1}} \Delta x_{1,t-q{_1}} + \cdots + \sum_{q_k=0}^{Q_k} \beta_{kq_k} \Delta x_{k,t-q{_k}} + \epsilon_t
-\end{split}
-\end{align} Here, the dependent variable---which can appear in either level form or first-differences---is a function of a constant, $\alpha_0$, a deterministic linear trend $T$, up to $P$ lags of the dependent variable and $L$ lags of the independent variables appearing in level form, up to $M$ lags of the first-differenced dependent variable and up to $Q$ lags of any first-differenced independent variables, and an independent and identically distributed error term $\epsilon_t$. Note that $L$ and $Q$ may differ for each of the included $k$ regressors. Users should use a combination of theory, information criteria, unit-root, cointegration (if applicable) and residual-based tests to arrive at a more restrictive specification of Equation \ref{eq:ardlgeneral} [@de2008taking;@philips2016have;@wilkins2018lag].
-
[…]
