+++ joss-doc/joss-graph-producer.R
@@ -1,339 +1,0 @@
-# Series simulator to make pretty pictures from dynardl
-library(dynamac)
-
-set.seed(1)
-x.error <- rnorm(500, 0, 0.5)
-x.full <- x.error
-y.error <- rnorm(500, 0, 2)
-y.full <- y.error
-y.diff.full <- rep(NA, length(y.full))
-phi <- -0.8 # coef on ldv
-beta.lx <- 1 # coef on l.x
-beta.diff.x <- -2 # coef on x
-
-
-for(i in 2:length(x.full)) {
-	x.full[i] <- x.full[i - 1] + x.error[i]
-	y.diff.full[i] <- phi*y.full[i - 1] + beta.diff.x*(x.full[i] - x.full[i - 1]) + beta.lx*x.full[i - 1] + y.error[i] # rnorm(1, 0, 1) + const*rnorm(1, 1, 0.25) +
-	y.full[i] <- y.full[i - 1] + y.diff.full[i]
-}
-
-# Trim the burn in
-x <- x.full[101:500]
-y.diff <- y.diff.full[101:500]
-y <- y.full[101:500]
-
-set.seed(1)
-model.ec <- dynardl(y ~ x, lags = list("x" = c(1)), diffs = c("x"), 
-	simulate = TRUE, fullsims = TRUE, 
-	shockvar = "x",
-	sims = 10000, range = 20,
-	ec = TRUE)
-	
-summary(model.ec)
-
-# This is NEW Figure 1.
-dynardl.all.plots(model.ec, bw = TRUE, tol = 0.05)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-beta.x <- 2 # coef on x
-beta.lx <- -3 # coef on l.x
-phi <- 0.5 # coef on ldv
-const <- 2.5 # constant needed to have a value of mean.y
-mean.y <- const/(1-phi) # steady-state of Y
-
-set.seed(100)
-x.full <- rnorm(500)
-y.full <- rep(NA, length(x.full))
-
-y.full[1] <- const/1-phi
-
-for(i in 2:length(x.full)) {
-#	y.full[i] <- const*rnorm(1, 1, 0.05) + phi*y.full[i - 1] + beta.x*x.full[i] + beta.lx*x.full[i - 1] + rnorm(1, 0, 0.05)
-  y.full[i] <- const + phi*y.full[i - 1] + beta.x*x.full[i] + beta.lx*x.full[i - 1] + rnorm(1, 0, 2)
-  
-}
-
-# Trim the burn in
-x <- x.full[101:500]
-y <- y.full[101:500]
-
-par(mfrow = c(2, 1))
-plot(y)
-plot(x)
-dev.off()
-
-set.seed(1)
-model <- dynardl(y ~ x, lags = list("x" = c(1)), levels = c("x"), 
-	simulate = TRUE, fullsims = TRUE, 
-	shockvar = "x",
-	sims = 10000, range = 20,
-	ec = FALSE)
-	
-summary(model)
-
-# This is Figure 1.
-dynardl.all.plots(model, bw = TRUE, tol = 0.05)
-
-
-
-
-
-
-
-
-
-############ EC
-# set.seed(1)
-# const <- 2 # constant needed to have a value of mean.y
-# x.full <- rep(NA, 500)
-# y.diff.full <- rep(NA, length(x.full))
-# y.full <- rep(NA, length(x.full))
-# y.full[1] <- 1
-# x.full[1] <- 15
[…]
