### 2020-09-28 #2528 @andyphilips
Thanks @sbenthall!

1. We updated the paper accordingly with your suggestion of adding an equation early on, did a final proofread, and uploaded.
2. We archived the current version of the program on Zenodo. DOI is https://doi.org/10.5281/zenodo.4057108. The current version of the software is 0.1.11.
