+++ README.md
@@ -4,8 +4,7 @@
 `dynamac` performs dynamic simulation and testing for single-equation ARDL models in R and Stata. Link to the webpage can be found [here](https://andyphilips.github.io/dynamac/).
 
 ## Installation (R)
-[![Build Status](https://travis-ci.com/andyphilips/dynamac.svg?branch=master)](https://travis-ci.com/andyphilips/dynamac) [![Build status](https://ci.appveyor.com/api/projects/status/o8h5gdh5cuah359y?svg=true)](https://ci.appveyor.com/project/andyphilips/dynamac) ![](https://www.r-pkg.org/badges/version/dynamac) [![codecov](https://codecov.io/gh/andyphilips/dynamac/branch/master/graph/badge.svg)](https://codecov.io/gh/andyphilips/dynamac) [![metacran downloads](https://cranlogs.r-pkg.org/badges/dynamac)](https://cran.r-project.org/package=dynamac) [![doi](https://img.shields.io/badge/DOI:-10.32614%2FRJ--2018--076-blueviolet.svg)](https://doi.org/10.32614/RJ-2018-076)
-
+[![Build Status](https://travis-ci.com/andyphilips/dynamac.svg?branch=master)](https://travis-ci.com/andyphilips/dynamac) [![Build status](https://ci.appveyor.com/api/projects/status/o8h5gdh5cuah359y?svg=true)](https://ci.appveyor.com/project/andyphilips/dynamac) ![](https://www.r-pkg.org/badges/version/dynamac) [![codecov](https://codecov.io/gh/andyphilips/dynamac/branch/master/graph/badge.svg)](https://codecov.io/gh/andyphilips/dynamac) [![metacran downloads](https://cranlogs.r-pkg.org/badges/dynamac)](https://cran.r-project.org/package=dynamac) [![doi](https://img.shields.io/badge/DOI:-10.32614%2FRJ--2018--076-blueviolet.svg)](https://doi.org/10.32614/RJ-2018-076) [![DOI](https://zenodo.org/badge/129036070.svg)](https://zenodo.org/badge/latestdoi/129036070)
 
 `dynamac` in R can easily be installed from its CRAN repository:
 ```
