### 2020-10-06 #2528 @kthyng
@andyphilips In terms of your paper: They are [meant to be](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) 250–1000 words and a quick word count of your paper (including references) was over 2000. So, without references it's not so long, but still is longer than we prefer. Please consider if there is a part that you can leave out.

Also, I saw two references in which "China" was not capitalized. You can force capitalization in Bibtex by putting {} around the word for which you want to preserve capitalization, like {China}. Can you go through your references in detail to check them? I can help in more detail if needed too.
