### 2020-10-08 c966592
made more word cuts
