### 2020-10-08 #2528 @andyphilips
Hi @kthyng, thanks for this. We have now done the following:
* Updated metadata on Zenodo to match title and author order
* Fixed capitalization issues in the references
* Performed a substantial word cut. We are now at 1280 words, including references but not including figures and Eq 1. I've updated the .md accordingly
