### 2020-10-09 #2528 @kthyng
@andyphilips I found one more capitalization to fix and did in this PR [#8](https://github.com/andyphilips/dynamac/pull/8) if you want to merge that on in.

Regarding removing content from your paper, I should have specified that you can certainly keep the content, but maybe move it to your docs. The paper itself is meant to stay brief but any useful information you can can go into the repo somewhere.
