### 2020-10-09 #2528 @andyphilips
Thanks @kthyng for fixing that! I just merged pull request #8 into master.
